Enable CircleCI
We are using two languages (Java and Go) and CircleCI doesn't have to have
first-class support for such a thing. The solution adopted in this commit is to
use the Java environment and fetch the Go using the Travis gimme script.

For simplicity the two gradle-plugin and the lib are build as part of the job.
The Maven packages are saved as artifacts. The JUnit results are also stored
using the `store_test_results` CircleCI command.